Move JS modules into src/ and update imports
Move 20 JavaScript engine and UI modules from the project root into a new src/ directory, add src/README.md documenting the module map and import rules, and update entry-point imports accordingly. Updated index.html imports (5 paths) and tests/battle.test.js imports (3 paths) to point to ./src/*. No logic changes to modules themselves — internal imports remain relative and unchanged. CHANGELOG.md updated to note the reorganization and reasoning; tests reported unchanged (97/97) in the changelog entry.
