+++ src/applyKnockback.js
