+++ src/audioManager.js
