+++ src/emanations.js
